auth_service/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asn1crypto==1.5.1
bcrypt==4.2.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
fastapi==0.115.6
filelock==3.16.1
h11==0.14.0
idna==3.10
packaging==24.2
passlib==1.7.4
platformdirs==4.3.6
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
PyJWT==2.10.1
pyOpenSSL==24.3.0
pytz==2024.2
requests==2.32.3
sniffio==1.3.1
snowflake-connector-python==3.12.4
sortedcontainers==2.4.0
starlette==0.41.3
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
